ajustando teste de datasource

+++ flutter_project/lib/src/infrastructure/data_sources/products_data_source.dart
@@ -1,5 +1,6 @@
 export 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_project/src/infrastructure/data_sources/i_products_data_source.dart';
 import 'package:flutter_project/src/infrastructure/dtos/product_dto.dart';
 
@@ -10,11 +11,30 @@ class ProductsDataSource implements IProductDataSource {
 
   @override
   Future delete(String id) async {
+    final exists = await _firestore.collection('products').doc(id).get();
+
+    if (exists.data() == null) {
+      throw PlatformException(
+        code: '404',
+        message: 'NOT_FOUND',
+      );
+    }
+
     await _firestore.collection('products').doc(id).delete();
   }
 
   @override
   Future update(ProductDto product) async {
+    final exists =
+        await _firestore.collection('products').doc(product.id).get();
+
+    if (exists.data() == null) {
+      throw PlatformException(
+        code: '404',
+        message: 'NOT_FOUND',
+      );
+    }
+
     await _firestore
         .collection('products')
         .doc(product.id)
